Retire React Bootstrap
React Bootstrap is currently migrating to be based on Bootstrap 4 which
is a change that we're not currently ready to make. Although I believe
that security updates might be issued for the library all the
documentation is being switched over to newer versions which makes it
difficult to understand how to structure the old-style components.

If you think some components of this library are particularly useful we
should start thinking about having some standard POP components instead
that implement our hybrid of Bootstrap 3 plus our Styleguide.